inherit wip

+++ tools/file_management.py
@@ -3,6 +3,7 @@
 import sys
 import os
 import subprocess
+import shlex
 import json
 from pathlib import Path
 
@@ -30,19 +31,19 @@ m_custom_json_dirname = "custom_json"
 m_custom_json_filename = "custom_json.json"
 m_module_log_filename = raw_code.module_log_fname
 
-def create_empty_module(module_name, initial_dir):
+def create_empty_module(module_name, ini_dir):
 	#change to modules dir
 	print()
 	os.chdir(modules_dir)
 
-    #create module file and chmod
+	#create module file and chmod
 	path_and_fn = Path(modules_dir) / module_name
 	path_and_fn.touch()
 	print("-"*40)
 	print(f"File created: {path_and_fn}")
 	
 	#return back to initial dir
-	os.chdir(initial_dir)
+	os.chdir(ini_dir)
 
 def get_unlabelled_dict(labelled_list, target_label):
 	for pos, ele in enumerate(labelled_list):
@@ -106,16 +107,16 @@ def write_to_file(filename,path,content, access_mode):
 				json.dump(content, target)
 				print(); print(f"\033[43mResponse saved to JSON file {full_path}\033[0m");
 
-def get_dict_value_save_to_file(gpt_response, initial_dir, filename, header=""):
-    print(); print("-"*40)
-    code = ut.get_response_value_for_key(gpt_response,raw_code.module_name.split(".")[0])
-    print("Code:"); print(); print(code)
-    destination_full_name = os.path.join(modules_dir, filename)
-    if not os.path.exists(destination_full_name):
-        create_empty_module(filename, initial_dir)
-    #add bash to script
-    code = header + code
-    write_to_file(filename,modules_dir,code, "w")
+def get_dict_value_save_to_file(gpt_response, ini_dir, filename, header=""):
+	print(); print("-"*40)
+	code = ut.get_response_value_for_key(gpt_response,raw_code.module_name.split(".")[0])
+	print("Code:"); print(); print(code)
+	destination_full_name = os.path.join(modules_dir, filename)
+	if not os.path.exists(destination_full_name):
+		create_empty_module(filename, ini_dir)
+	#add bash to script
+	code = header + code
+	write_to_file(filename,modules_dir,code, "w")
 
 def version_file(path_original_fn, original_fn, path_dest_fn):
 	original_full_path_fn = os.path.join(path_original_fn, original_fn)
@@ -133,51 +134,65 @@ def version_file(path_original_fn, original_fn, path_dest_fn):
 			counter += 1
 		else:
 			break
-			
+
 	command = f"cp {original_full_path_fn} {destination_full_name}"
-	subprocess.run(command, shell=True)
-	print(); print("-"*40);print()
-	print(f"\033[43mVersion saved: {command}\033[0m")
+	print(); print("-" * 40); print()
+	try:
+		#comm_list = shlex.split(command)
+		subprocess.run(command, check=True)
+		print(f"\033[43mVersion saved: {destination_full_name}\033[0m")
+	except FileNotFoundError as e:
+		print(f"\033[43mFile {original_full_path_fn} not available to version it\033[0m")
+	except Exception as e:
+		print(f"\033[43mException versioning (copy) {original_full_path_fn} to {destination_full_name}: {e}\033[0m")
 
-def read_file_stored_to_buffer(filename,path):
-		full_path=os.path.join(path, filename)
-		path_to_file = Path(path,filename)
-		ext = path_to_file.name.split(".")[1]
+def read_file_stored_to_buffer(filename, path):
+	full_path = os.path.join(path, filename)
+	path_to_file = Path(path, filename)
+	# if not os.path.isfile(path_to_file):
+	# 	print(f"\033[1;31m[ERROR]\033[0m Cannot Find Script File {path_to_file}\033[0m")
+	ext = path_to_file.name.split(".")[1]
 
-		if ext != "json":
-			return Path(full_path).read_text()
+	if ext != "json":
+		return Path(full_path).read_text()
 
-		#thanks to https://gist.github.com/tomschr/86a8c6f52b81e35ac4723fef8435ec43
-		with Path(full_path).open(encoding="UTF-8") as buff:
-			return json.load(buff)
+	#thanks to https://gist.github.com/tomschr/86a8c6f52b81e35ac4723fef8435ec43
+	with Path(full_path).open(encoding="UTF-8") as buff:
+		return json.load(buff)
 
 def print_json_on_screen(json_data):
-        print(json.dumps(json_data, indent=2, separators=(',', ':')))
+	print(json.dumps(json_data, indent=2, separators=(',', ':')))
 
 def insert_script_in_json(a_script):
-    json_dict = {"module":a_script}
-    #return json obj
-    json_str = json.dumps(json_dict)
-    a = json.loads(json_str)
-    return json.loads(json_str)
+	json_dict = {"module":a_script}
+	#return json obj
+	json_str = json.dumps(json_dict)
+	a = json.loads(json_str)
+	return json.loads(json_str)
 
 def delete_all_dir_files(target_dir):
-		#del all project module files
+	#del all project module files
+	if os.path.exists(target_dir):
 		with os.scandir(target_dir) as this_dir:
 			for file_name in this_dir:
 				if file_name.is_file():
 					os.remove(os.path.join(modules_dir, file_name))
 
 def create_dir(target_dir):
-        #os.makedirs(target_dir, exist_ok=True)
-        target_dir.mkdir(parents=False, exist_ok=True)
+	#os.makedirs(target_dir, exist_ok=True)
+	target_dir.mkdir(parents=False, exist_ok=True)
 
 def validate_filepath(full_path_to_script):
 	if not os.path.isfile(full_path_to_script):
-	    print(f"\033[1;31m[ERROR]\033[0m Cannot Find File {full_path_to_script}\033[0m")
-	    return False
-	else: 
-	    print("File Found.")
-	    return True
+		print(f"\033[1;31m[ERROR]\033[0m Cannot Find File {full_path_to_script}\033[0m")
+		return False
+	else:
+		print("File Found.")
+		if os.access(full_path_to_script, os.R_OK):
+			return True
+		else:
+			print(f"\033[1;31m[ERROR]\033[0m File {full_path_to_script} not readable\033[0m")
+			return False
+
